docs(vetting): correct the 3D NGLDM report's file-inventory claim, and widen the generator's shift
The report's closing paragraph said tests/vetting/matrix/ngldm.md and
tests/vetting/benchmarks.md "do not exist, and exist for no family but GLCM". benchmarks.md
does exist and is where bench_ut57_3d -- the benchmark all 19 3NGLDM rows cite -- is defined,
and matrix/ holds fifteen files including the ngldm3d.md this report links two sections
earlier. Only the 2D family's matrix/ngldm.md is genuinely absent, so that is what the
paragraph now says.

nyxus_grey_levels() subtracted the volume minimum in the array's on-disk dtype before
widening to uint32. NIfTI datatype 4 is int16, so a volume spanning more than 32767 would
wrap there and the levels handed to MIRP would stop reproducing what Nyxus bins to -- a
silent wrong answer in the run that the samelevels comparison rests on. The subtraction now
happens in int64. The current phantom is float32 and never reached it.

Documentation and generator only; no feature value moves. gen_ngldm3d_mirp.py re-run: 34
verified, 0 failed, 0 unproducible, 0 unquoted.

+++ tests/vetting/oracles/gen_ngldm3d_mirp.py
@@ -141,7 +141,9 @@ def nyxus_grey_levels(inten, mask):
     volume minimum, and `to_grayscale(i, 0, ROI max, 64)` truncates i / (ROI max) * 64. Handing the
     result to MIRP with the discretisation switched off is what makes the samelevels run comparable.
     """
-    shifted = (inten - inten.min()).astype(np.uint32)
+    # widen before subtracting: NIfTI datatype 4 is int16, and a volume spanning more than 32767
+    # would wrap in the file's own dtype and stop reproducing Nyxus' binning
+    shifted = (inten.astype(np.int64) - int(inten.min())).astype(np.uint32)
     roi_max = float(shifted[mask].max())
     levels = (shifted.astype(np.float64) / roi_max * NBINS).astype(np.uint32)
     return np.where(mask, levels, 0).astype(np.float64)
